+++ packages/app/lib/i18n/gen/strings.g.dart
@@ -4,9 +4,9 @@
 /// To regenerate, run: `dart run slang`
 ///
 /// Locales: 3
-/// Strings: 1876 (625 per locale)
+/// Strings: 1873 (624 per locale)
 ///
-/// Built on 2026-05-31 at 20:33 UTC
+/// Built on 2026-05-31 at 22:44 UTC
 
 // coverage:ignore-file
 // ignore_for_file: type=lint, unused_import